Some directory refactors keeping in mind future work
