+++ src/managers/resources/bms.rs
@@ -6,19 +6,17 @@ use tokio::sync::mpsc::Receiver;
 use tokio::sync::mpsc::Sender;
 use tokio::sync::mpsc::channel;
 
-use super::MAX_MPSC_CHANNEL_BUFFER;
-use super::Manager;
-use super::ManagerChannelData;
-use super::Resource;
-use super::ResourceManager;
-use super::Responder;
-use super::TASK_SUCCESS;
-use crate::parse_channel_data;
+use crate::managers::MAX_MPSC_CHANNEL_BUFFER;
+use crate::managers::Manager;
+use crate::managers::ManagerChannelData;
+use crate::managers::Resource;
+use crate::managers::ResourceManager;
+use crate::managers::Responder;
+use crate::managers::TASK_SUCCESS;
 use crate::request::TaskData::BmsData;
 use crate::sgcp;
 use crate::sgcp::bms::*;
 use crate::todo;
-use crate::verify_channel_data;
 
 /// Represents a BMS resource
 pub struct Bms {
