+++ src/managers/resources/emg.rs
@@ -8,19 +8,17 @@ use tokio::sync::mpsc::Receiver;
 use tokio::sync::mpsc::Sender;
 use tokio::sync::mpsc::channel;
 
-use super::MAX_MPSC_CHANNEL_BUFFER;
-use super::Manager;
-use super::ManagerChannelData;
-use super::Resource;
-use super::ResourceManager;
-use super::Responder;
-use super::TASK_SUCCESS;
-use crate::parse_channel_data;
+use crate::managers::MAX_MPSC_CHANNEL_BUFFER;
+use crate::managers::Manager;
+use crate::managers::ManagerChannelData;
+use crate::managers::Resource;
+use crate::managers::ResourceManager;
+use crate::managers::Responder;
+use crate::managers::TASK_SUCCESS;
 use crate::request::TaskData::EmgData;
 use crate::sgcp;
 use crate::sgcp::emg::*;
 use crate::todo;
-use crate::verify_channel_data;
 
 /// Represents an EMG resource
 pub struct Emg {
