+++ src/managers/resources/mod.rs
@@ -1,0 +1,3 @@
+pub mod bms;
+pub mod emg;
+pub mod maestro;
